Change return type of FrameDescription::GetFrameSize to avoid unneeded type casts. Review URL: http://codereview.chromium.org/7282033
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@8500 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arm/deoptimizer-arm.cc
@@ -550,7 +550,7 @@ void Deoptimizer::FillInputFrame(Address tos, JavaScriptFrame* frame) {
   }
 
   // Fill the frame content from the actual data on the frame.
-  for (intptr_t i = 0; i < input_->GetFrameSize(); i += kPointerSize) {
+  for (unsigned i = 0; i < input_->GetFrameSize(); i += kPointerSize) {
     input_->SetFrameSlot(i, Memory::uint32_at(tos + i));
   }
 }
